style: reformat code base

+++ app/src/main/java/com/example/btlthadr/Activity/CartActivity.java
@@ -2,10 +2,14 @@ package com.example.btlthadr.Activity;
 
 import android.view.View;
 import android.widget.LinearLayout;
+
 import androidx.appcompat.app.AppCompatActivity;
+
 import android.os.Bundle;
+
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
 import com.example.btlthadr.Adapter.CartAdapter;
 import com.example.btlthadr.Helper.ChangeNumberItemsListener;
 import com.example.btlthadr.Helper.ManagmentCart;
@@ -19,14 +23,13 @@ public class CartActivity extends BaseActivity {
     private double tax;
 
 
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        binding=ActivityCartBinding.inflate(getLayoutInflater());
+        binding = ActivityCartBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
-        managmentCart=new ManagmentCart(this);
+        managmentCart = new ManagmentCart(this);
         setVariable();
         calculateCart();
         initList();
@@ -34,36 +37,35 @@ public class CartActivity extends BaseActivity {
     }
 
     private void initList() {
-        if(managmentCart.getListCart().isEmpty()){
+        if (managmentCart.getListCart().isEmpty()) {
             binding.emptyTxt.setVisibility(View.VISIBLE);
             binding.scrollViewCart.setVisibility(View.GONE);
-        }
-        else{
+        } else {
             binding.emptyTxt.setVisibility(View.GONE);
             binding.scrollViewCart.setVisibility(View.VISIBLE);
 
         }
-        LinearLayoutManager linearLayoutManager=new LinearLayoutManager(this,LinearLayoutManager.VERTICAL,false);
+        LinearLayoutManager linearLayoutManager = new LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false);
         binding.cardView.setLayoutManager(linearLayoutManager);
-        adapter=new CartAdapter(managmentCart.getListCart(), this, () -> calculateCart());
+        adapter = new CartAdapter(managmentCart.getListCart(), this, () -> calculateCart());
         binding.cardView.setAdapter(adapter);
     }
 
     private void calculateCart() {
-        double percentTax=0.02;
-        double delivery=10;
-        tax= (double) Math.round(managmentCart.getTotalFee() * percentTax * 100.0) /100;
-        double total=Math.round((managmentCart.getTotalFee()+tax+delivery)*100)/100;
-        double itemTotal=Math.round(managmentCart.getTotalFee()*100)/100;
-        binding.totalFeeTxt.setText("$"+itemTotal);
-        binding.taxTxt.setText("$"+tax);
-        binding.deliveryTxt.setText("$"+delivery);
-        binding.totalTxt.setText("$"+total);
+        double percentTax = 0.02;
+        double delivery = 10;
+        tax = (double) Math.round(managmentCart.getTotalFee() * percentTax * 100.0) / 100;
+        double total = (double) Math.round((managmentCart.getTotalFee() + tax + delivery) * 100) / 100;
+        double itemTotal = (double) Math.round(managmentCart.getTotalFee() * 100) / 100;
+        binding.totalFeeTxt.setText("$" + itemTotal);
+        binding.taxTxt.setText("$" + tax);
+        binding.deliveryTxt.setText("$" + delivery);
+        binding.totalTxt.setText("$" + total);
 
 
     }
 
-    public void setVariable(){
+    public void setVariable() {
         binding.backBtn.setOnClickListener(v -> finish());
     }
 }
